split : CueSheetParser 코드 분리
